added FC read-in functionality:
If one wants to repeat a calculation with different parameters of the simulation but the same potentials, i.e. the same Franck-Condon integrals,
this new functionality offers the possibility to use pre-calculated FC integrals as additional input.
This is triggered by providing a second input file when calling nuclear_dyn.
The second file is thought to be a reduced copy of eldest.out from a previous calculation with the same potentials;
it may start with an arbitrary number of lines (including zero) preceding the gs-fin FC integrals
PROVIDED that their first word is not numeric.
Then the gs-fin integrals shall follow, with the first word indicating the gs quantum number and the last word being the integral value.
At least one line beginning with a non-numeric word then separates the gs-fin integrals from the res-fin integrals.
These shall have the structure as described for the gs-fin integrals.
There shall be no dividing line between different res quantum numbers.
The first line not beginning with a numeric word then indicates the endpoint for the read-in routine;
all lines thereafter will be ignored, regardless of their first word.

Note: The gs-res FC integrals are NOT read in by this routine and shall not be present in the second input file.
Since gs and res are assumed to be only a small number of bound states, the integrals will be directly calculated.
It was decided against the possibility to read them in as well
because these gs-res integrals are (calculated and) printed at a different location in nuclear_dyn than gs-fin and res-fin.
For the sake of keeping the code simple and due to the low computational effort of calculating gs-res, they are not included in the read-in routine.